test(integration tests): split integration testing into 3 files
