+++ test/integration2.test.js
@@ -1,0 +1,231 @@
+/* eslint-disable no-undefined */
+/* globals
+  describe
+  global
+  test
+*/
+
+import 'jsdom-worker';
+
+import 'jquery-ui-bundle';
+import 'materialize-css';
+import { getIdFromSensorName, getIdFromStarName, getSat, getSatPosOnly, satSet } from '@app/js/satSet/satSet.js';
+import { uiInput, uiManager } from '@app/js/uiManager/uiManager.js';
+import { Camera } from '@app/js/cameraManager/camera.js';
+import { ColorSchemeFactory as ColorScheme } from '@app/js/colorManager/color-scheme-factory.js';
+import { GroupFactory } from '@app/js/groupsManager/group-factory.js';
+import { LineFactory } from '@app/js/drawManager/sceneManager/sceneManager.js';
+import { drawManager } from '@app/js/drawManager/drawManager.js';
+import { objectManager } from '@app/js/objectManager/objectManager.js';
+import { orbitManager } from '@app/js/orbitManager/orbitManager.js';
+import { photoManager } from '@app/js/photoManager/photoManager.js';
+import { sMM } from '@app/js/uiManager/sideMenuManager.js';
+import { satellite } from '@app/js/lib/lookangles.js';
+import { searchBox } from '@app/js/uiManager/search-box.js';
+import { sensorManager } from '@app/js/sensorManager/sensorManager.js';
+import { settingsManager } from '@app/js/settingsManager/settingsManager.js';
+import { starManager } from '@app/js/starManager/starManager.js';
+import { timeManager } from '@app/js/timeManager/timeManager.js';
+
+// eslint-disable-next-line sort-imports
+import $ from 'jquery';
+import { missileManager } from '@app/js/missileManager/missileManager.js';
+
+// eslint-disable-next-line sort-imports
+import { eventFire, setup } from './setup.js';
+
+setup.init();
+
+const exampleSat = {
+  C: 'US',
+  LS: 'AFETR',
+  LV: 'U',
+  ON: 'VANGUARD 1',
+  OT: 1,
+  R: '0.1220',
+  SCC_NUM: '00005',
+  TLE1: '1     5U 58002B   21107.45725112 -.00000113  00000-0 -16194-3 0  9999',
+  TLE2: '2     5  34.2637  11.6832 1848228 280.4329  59.4145 10.84843191238363',
+  active: true,
+  apogee: 3845.1282721399293,
+  argPe: 4.894477435916007,
+  eccentricity: 0.1848228,
+  id: 0,
+  inclination: 0.5980143789155811,
+  intlDes: '1958-002B',
+  meanMotion: 10.843102290386977,
+  perigee: 657.8610581463026,
+  period: 132.80332154356245,
+  position: {
+    x: 4000,
+    y: 4000,
+    z: 4000,
+  },
+  velocity: {
+    x: 7,
+    y: 7,
+    z: 7,
+    total: 14,
+  },
+  raan: 0.2039103071690015,
+  semiMajorAxis: 8622.494665143116,
+  semiMinorAxis: 8473.945136538932,
+  getAltitude: () => 100,
+  getDirection: () => 'N',
+  isInSun: () => true,
+  getTEARR: () => {
+    const currentTEARR = {
+      lat: 0.1,
+      lon: 0.1,
+      alt: 50000,
+      az: 0,
+      el: 0,
+      rng: 0,
+    };
+    satellite.setTEARR(currentTEARR);
+    return currentTEARR;
+  },
+};
+
+describe('Integration Testing', () => {
+  var lineManager, groupsManager, cameraManager, gl, dotsManager;
+
+  test('Setup Test Environment', async () => {
+    await timeManager.init();
+    settingsManager.loadStr('dots');
+    uiManager.mobileManager.init();
+    cameraManager = new Camera();
+    // We need to know if we are on a small screen before starting webgl
+    gl = await drawManager.glInit();
+
+    // NOTE: Jest fails with webgl2 so we use webgl1 during testing
+    // This means we need to mock some of the webgl2 code
+    gl = global.mocks.glMock;
+
+    window.addEventListener('resize', drawManager.resizeCanvas);
+    drawManager.loadScene(gl);
+    dotsManager = await drawManager.createDotsManager();
+    satSet.init(gl, dotsManager, cameraManager);
+    objectManager.init(sensorManager);
+    await ColorScheme.init(gl, cameraManager, timeManager, sensorManager, objectManager, satSet, satellite, settingsManager);
+    drawManager.selectSatManager.init(ColorScheme.group, sensorManager, satSet, objectManager, sMM, timeManager);
+    await satSet.loadCatalog(); // Needs Object Manager and gl first
+    const satCruncher = satSet.satCruncher;
+
+    dotsManager.setupPickingBuffer(satSet.satData);
+    satSet.setColorScheme(ColorScheme.default, true);
+
+    groupsManager = new GroupFactory(satSet, ColorScheme, settingsManager);
+    await orbitManager.init(gl, cameraManager, groupsManager);
+    searchBox.init(satSet, groupsManager, orbitManager, dotsManager);
+    lineManager = new LineFactory(gl, orbitManager.shader, getIdFromSensorName, getIdFromStarName, getSat, getSatPosOnly);
+    starManager.init(lineManager, getIdFromStarName);
+    uiManager.init(cameraManager, lineManager, starManager, groupsManager, satSet, orbitManager, groupsManager, ColorScheme);
+    await satellite.initLookangles(satSet, satCruncher, sensorManager, groupsManager);
+    dotsManager.updateSizeBuffer(satSet.satData);
+    // await radarDataManager.init(sensorManager, satSet, satCruncher, satellite);
+    satSet.setColorScheme(settingsManager.currentColorScheme); // force color recalc
+    objectManager.satLinkManager.idToSatnum(satSet);
+
+    uiInput.init(cameraManager, objectManager, satellite, satSet, lineManager, sensorManager, starManager, ColorScheme, satCruncher, uiManager, drawManager, dotsManager);
+
+    drawManager.init(groupsManager, uiInput, starManager, satellite, ColorScheme, cameraManager, objectManager, orbitManager, sensorManager, uiManager, lineManager, dotsManager);
+
+    // Now that everything is loaded, start rendering to thg canvas
+    drawManager.drawLoop();
+
+    // UI Changes after everything starts -- DO NOT RUN THIS EARLY IT HIDES THE CANVAS
+    uiManager.postStart();
+    photoManager.init(cameraManager, satSet, timeManager, uiManager, drawManager.selectSatManager);
+
+    // Make testing go faster
+    satellite.lookanglesLength = 0.5;
+  });
+
+  test('Missile Manager Functional', () => {
+    let sim = 'simulation/Russia2USA.json';
+    missileManager.MassRaidPre(new Date(), sim);
+
+    let a = 101 - 100;
+    let b = 500 - 0;
+    missileManager.Missile(0, 0, 41, -71, 3, satSet.missileSats - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.clearMissiles();
+
+    missileManager.Missile(0, 0, 100, -71, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], null, null, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, -100, -71, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, 40, -501, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, 40, 501, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+
+    missileManager.Missile(100, 0, 0, -71, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(-100, 0, 0, -71, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 300, 40, 1, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, -300, 40, 1, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, 40, 1, 15, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, 40, 1, 15.1, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], 30, 2.9, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+
+    missileManager.Missile(0, 0, 1, 1, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], null, null, 0.07, missileManager.UsaICBM[a * 4 + 3], 'United States', 13000);
+    missileManager.Missile(0, 0, 0, 180, 3, 500 - b, new Date(), missileManager.UsaICBM[a * 4 + 2], null, null, 0.07, 1000, 'United States', 13000);
+
+    $('#ms-type').val(0);
+    $('#ms-attacker').val(101);
+    $('#ms-lat-lau').val(0);
+    $('#ms-lon-lau').val(0);
+    $('#ms-target').val('-1');
+    $('#ms-lat').val(40);
+    $('#ms-lon').val(180);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(201);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(301);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(401);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(501);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(601);
+    eventFire('missile', 'submit');
+    $('#ms-attacker').val(701);
+    eventFire('missile', 'submit');
+
+    $('#ms-type').val(1);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(2);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(3);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(4);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(5);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(6);
+    eventFire('missile', 'submit');
+    $('#ms-type').val(7);
+    eventFire('missile', 'submit');
+
+    $('#ms-lat-lau').val('a');
+    eventFire('missile', 'submit');
+    $('#ms-lon-lau').val('a');
+    $('#ms-lat').val('a');
+    eventFire('missile', 'submit');
+    $('#ms-lon').val('a');
+    eventFire('missile', 'submit');
+  });
+
+  test('photoManager Functional', () => {
+    satSet.getSatFromObjNum = () => exampleSat;
+    satSet.getSatExtraOnly = () => exampleSat;
+
+    const req = {
+      response:
+        '[{"identifier":"20210530010437","caption":"This image was taken by NASA\'s EPIC camera onboard the NOAA DSCOVR spacecraft","image":"epic_1b_20210530010437","version":"03","centroid_coordinates":{"lat":21.291504,"lon":165.9375},"dscovr_j2000_position":{"x":521173.554246,"y":1386579.317085,"z":576553.775},"lunar_j2000_position":{"x":173282.377382,"y":-289823.648477,"z":-153187.465575},"sun_j2000_position":{"x":55408994.591638,"y":129514078.915438,"z":56143391.809486},"attitude_quaternions":{"q0":0.5293,"q1":-0.55936,"q2":-0.5925,"q3":0.23644},"date":"2021-05-30 00:59:48","coords":{"centroid_coordinates":{"lat":21.291504,"lon":165.9375},"dscovr_j2000_position":{"x":521173.554246,"y":1386579.317085,"z":576553.775},"lunar_j2000_position":{"x":173282.377382,"y":-289823.648477,"z":-153187.465575},"sun_j2000_position":{"x":55408994.591638,"y":129514078.915438,"z":56143391.809486},"attitude_quaternions":{"q0":0.5293,"q1":-0.55936,"q2":-0.5925,"q3":0.23644}}}]',
+      status: 200,
+    };
+
+    photoManager.dscovr();
+    photoManager.dscovrLoaded(req);
+    photoManager.goes1();
+    photoManager.himawari8();
+    photoManager.meteosat8();
+    photoManager.meteosat11();
+  });
+});
